requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain
langchain-openai
python-dotenv
google-api-python-client
lxml
httpx
beautifulsoup4
trafilatura
